backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
anthropic==0.42.0
PyJWT==2.10.1
numpy==2.2.1
python-multipart==0.0.20
httpx==0.28.1
aiofiles==24.1.0
web3==7.6.0

pytest==8.3.4
pytest-asyncio==0.25.0
